Enum woab::Error

pub enum Error {

    IoError(Error),
    FromUtf8Error(FromUtf8Error),
    XmlError(Error),
    GtkBoolError(BoolError),
    WidgetMissingInBuilder(String),
    IncorrectWidgetTypeInBuilder {
        widget_id: String,
        expected_type: Type,
        actual_type: Type,
    },
    NoSuchSignalError(String),
    IncorrectSignalParameterType {
        signal: String,
        index: usize,
        expected_type: Type,
        actual_type: Type,
    },
    SignalParameterIndexOutOfBound {
        signal: String,
        index: usize,
        num_parameters: usize,
    },
    IncorrectEventParameter {
        signal: String,
        expected_type: &'static str,
        actual_type: EventType,
    },
    IncorrectActionParameter {
        signal: String,
        expected_type: VariantType,
        actual_type: VariantType,
    },
    NotAllParametersExtracted {
        signal: String,
        num_parameters: usize,
        num_extracted: usize,
    },
    WakerPerished(WakerPerished),

}

Variants

IoError(Error)

FromUtf8Error(FromUtf8Error)

XmlError(Error)

GtkBoolError(BoolError)

WidgetMissingInBuilder(String)

When extracting widgets using BuilderConnector::widgets and one of the widgets is missing.

IncorrectWidgetTypeInBuilder

When extracting widgets using BuilderConnector::widgets and one of the widgets has the wrong type.

Fields of IncorrectWidgetTypeInBuilder

widget_id: Stringexpected_type: Typeactual_type: Type

NoSuchSignalError(String)

When a signal handler does not recognize the name of the signal routed to it.

IncorrectSignalParameterType

When a signal parameter has the wrong type.

Fields of IncorrectSignalParameterType

signal: Stringindex: usizeexpected_type: Typeactual_type: Type

SignalParameterIndexOutOfBound

When a signal has less parameters than what the handler expects.

Fields of SignalParameterIndexOutOfBound

signal: Stringindex: usizenum_parameters: usize

IncorrectEventParameter

When an event signal’s parameter is of the the wrong event type.

Fields of IncorrectEventParameter

signal: Stringexpected_type: &'static stractual_type: EventType

IncorrectActionParameter

When an action signal’s parameter is of the the wrong type.

Fields of IncorrectActionParameter

signal: Stringexpected_type: VariantTypeactual_type: VariantType

NotAllParametersExtracted

When a signal has more parameters than what the handler expects.

Fields of NotAllParametersExtracted

signal: Stringnum_parameters: usizenum_extracted: usize

WakerPerished(WakerPerished)