Function woab::run_dialog

pub async fn run_dialog(
    dialog: &impl Clone + DialogExt + GtkWindowExt + WidgetExt + 'static, 
    close_after: bool
) -> ResponseType

Run a GTK dialog and get its response.

Use this instead of gtk-rs’ dialog.run(), because dialog.run() runs a nested GTK loop which does not play nice with WoAB and because run_dialog is async and can let Actix continue running in the background.

GTK does not automatically close dialogs when the user responds to them, but run_dialog can do it if true is passed as the second argument.

let dialog_response = woab::run_dialog(
    &gtk::MessageDialog::new::<gtk::ApplicationWindow>(
        None,
        gtk::DialogFlags::all(),
        gtk::MessageType::Question,
        gtk::ButtonsType::YesNo,
        "What would it be?",
    ),
    true,
)
.await;
println!("User says {}", dialog_response);