Function woab::wake_from

pub async fn wake_from<T>(
    setup_dlg: impl FnOnce(Sender<T>)
) -> Result<T, WakerPerished>

Asynchronously wait for something to happen somewhere.

Accepts a closure that accepts a Sender. The closure must “plant” the sender somewhere - usually inside a signal - and once the sender’s do_send is called, wake_from will be woken and return the value passed to the sender.

Note that unless explicitly removed, any signal handler registered inside the closure would remain active afterwards - just like any other signal handler registered in GTK. To wake from a signal it is better to use woab::wake_from_signal.

let button1: gtk::Button;
let button2: gtk::Button;
let button_clicked = woab::wake_from(|tx| {
    button1.connect_clicked({
        let tx = tx.clone();
        move |_| {
            let _ = tx.try_send(1);
        }
    });
    button2.connect_clicked(move |_| {
        let _ = tx.try_send(2);
    });
}).await.unwrap();
println!("Clicked button {}", button_clicked);

Pitfalls

The sender’s do_send can only be called once, and will fail if called multiple times. Because GTK signals can be called multiple times, this cannot be guaranteed by the type system and the result of do_send needs to be swallowed if there is a possibility for it to be called multiple times.

The closure runs inside the same runtime that awaits the wake_from result. If that runtime is the Actix runtime and the closure needs to do anything that must not be ran from the Actix runtime, it’d need to either use spawn_outside inside the closure or use outside before wake_from is called.