Macro woab::params

params!() { /* proc-macro */ }

Helper macro for extracting signal parameters from woab::Signal.

let woab::params!(
    _,
    param1: String,
    param2,
) = msg.params()?; // `msg` is the `woab::Signal`

All the signal parameters must be matched against, but _ can be used for unneeded parameters. Parameters with types will be converted to that type, and untyped parameters will be &glib::Value.