Struct woab::BuilderConnector

pub struct BuilderConnector(_);

Context for utilizing a gtk::Builder and connecting it to he Actix world.

It wraps a gtk::Builder instance and provides methods to create actors that are connected to the widgets in that builder.

See BuilderFactory for usage example.

Implementations

impl BuilderConnector

pub fn get_object<W>(&self, id: &str) -> Result<W, Error> where
W: IsA<Object>,

Get a GTK object from the builder by id.

pub fn with_object<W>(self, id: &str, dlg: impl FnOnce(W)) -> Self where
W: IsA<Object>,

Fluent interface for doing something with a particular object from the builder.

This is useful for setting up the builder created widgets:

builder_factory.instantiate()
    .with_object("window", |window: gtk::ApplicationWindow| {
        window.show();
    })
    .connect_to(MyActor.start());

pub fn widgets<W>(&self) -> Result<W, <Builder as TryInto<W>>::Error> where
Builder: TryInto<W>,

Create a widgets struct who’s fields are mapped to the builder’s widgets.

pub fn connect_to(
self,
target: impl IntoGenerateRoutingGtkHandler
) -> BuilderConnectorWidgetsOnly

Route the builder’s signals to Actix.

bld.connect_to(target) will connect all the signals defined in the Glade XML to target, which can be an actix::Recipient<woab::Signal>, an actix::Addr<A> for A that handles woab::Signal, or a woab::NamespacedSignalRouter.

bld.connect_to((tag, target)) will connect all the signals to target and add a tag. The tag must be Clone, and will be sent along with the signal. The signal type will be woab::Signal<T> where T is the type of the tag.

builder_factory.instantiate().connect_to(MyActor.start());

pub fn connect_with<G: IntoGenerateRoutingGtkHandler>(
self,
dlg: impl FnOnce(&Self) -> G
) -> BuilderConnectorWidgetsOnly

Runs a closure and passes the result to connect_to.

This is mostly a convenience method. The closure receives the &BuilderConnector and can be use it to retrieve widgets from the builder instantiation and use them in the creation of the actor.

#[derive(woab::WidgetsFromBuilder)]
struct MyWidgets {
    my_button: gtk::Button,
    my_textfield: gtk::Entry,
}

struct MyActor {
    widgets: MyWidgets,
    my_window: gtk::Window, // for whatever reason not in `MyWidgets`
}
builder_factory.instantiate().connect_with(|bld| {
    MyActor {
        widgets: bld.widgets().unwrap(),
        my_window: bld.get_object("my_window").unwrap(),
    }
    .start()
});