Struct woab::BuilderConnectorWidgetsOnly

pub struct BuilderConnectorWidgetsOnly { /* fields omitted */ }

Degraded version of BuilderConnector that can only be used to get widgets.

After the BuilderConnector connects its signals, they cannot be connected again - so the BuilderConnector is consumed. But the widgets are still accessible with this object.

Implementations

impl BuilderConnectorWidgetsOnly

pub fn get_object<W>(&self, id: &str) -> Result<W, Error> where
W: IsA<Object>,

See BuilderConnector::get_object.

pub fn with_object<W>(&self, id: &str, dlg: impl FnOnce(W)) -> &Self where
W: IsA<Object>,

See BuilderConnector::with_object.

pub fn widgets<W>(&self) -> Result<W, <Builder as TryInto<W>>::Error> where
Builder: TryInto<W>,

See BuilderConnector::widgets.