Struct woab::BuilderSingalConnector

pub struct BuilderSingalConnector<S, T, I> where
    S: BuilderSignal,
    T: Clone,
    I: SignalsInhibit<S>,  { /* fields omitted */ }

Fluent interface for configuring the signal routing from GTK to Actix.

Typically created from BuilderSignal::connector, BuilderSingalConnector can be configured via fluent methods like tag and inhibit and then either:

Get passed to connect_signals that connects signals emitted from a freshly created GTK builder to a freshly created Actix actor.

Get converted to a SignalRouter via route_to to connect GTK signals - from builders or otherwise - to an Actix actor.

Implementations

impl<S, I> BuilderSingalConnector<S, (), I> where
S: BuilderSignal,
I: SignalsInhibit<S>,

pub fn tag<T: Clone>(self, tag: T) -> BuilderSingalConnector<S, (T,), I>

Add a tag for identifying the builder instance that generated the signal.

This is useful when you are using the version of connect_signals that does not also create an actor to connect multiple instantiations of the same set of widgets to a single actor, and need to tell the actor which set of widgets generated the signal.

The big example in BuilderConnector demonstrates how to use this.

impl<S, T> BuilderSingalConnector<S, T, ()> where
S: BuilderSignal,
T: SignalTransformer<S>,

pub fn inhibit<F: Clone + Fn(&S) -> Option<Inhibit>>(
self,
dlg: F
) -> BuilderSingalConnector<S, T, F>

Register a function for setting the return value of the signal.

GTK requires some signals to return a boolean value - true to "inhibit" and not let the signal pass up the inheritance to other handlers, and false to let it. This can be statically configured in the BuilderSignal derive macro with #[signal(inhibit = true) or #[signal(inhibit = false), but sometimes you need to decide based on the signal's parameters. In such cases you can use this function to register a closure. If the closure returns Some, the signal handler will return the value returned by the closure. If it returns None the signal handler will return the value configured in the macro.

enum MySignal {
    ButtonPress(gtk::Button, #[signal(event)] gdk::EventButton), // needs to return a boolean
    ButtonClick, // needs to return nothing
}

fn some_function_that_creates_stuff(factory: &Factories) {
    factory.window.instantiate().actor()
        .connect_signals(MySignal::connector().inhibit(|signal| match signal {
            MySignal::ButtonPress(_, event) => Some(gtk::Inhibit(event.get_button() == 3)), // inhibit on right click
            _ => None, // all other signals should use the default (static) setting
        }))
        .start(MyActor);
}

impl<S, T, I> BuilderSingalConnector<S, T, I> where
S: 'static,
S: BuilderSignal,
T: 'static,
T: SignalTransformer<S>,
I: 'static,
I: SignalsInhibit<S>,

pub fn route_to<A>(self, ctx: &mut A::Context) -> SignalRouter<S, I> where
A: Actor<Context = Context<A>>,
A: StreamHandler<T::Output>,

Create a SignalRouter that routes signals to the actor owning the supplied context.

This method only sets the target of the signal routing - the signal router object provides methods for connecting the signals from GTK. If they are not used no actual signals will be routed.