Trait woab::BuilderSignal

pub trait BuilderSignal: Sized + 'static {
    const SIGNALS: &'static [&'static str];

    fn bridge_signal(
        signal: &str, 
        tx: Sender<Self>, 
        inhibit_dlg: impl 'static + Fn(&Self) -> Option<Inhibit>
    ) -> Result<RawSignalCallback, Error>;

    fn connector() -> BuilderSingalConnector<Self, (), ()> { ... }
}

Represent a GTK signal that originates from a GTK builder. Refer to the corresponding derive.

Associated Constants

const SIGNALS: &'static [&'static str]

The list of signals supported by this BuilderSignal

Loading content...
Required methods

fn bridge_signal(
signal: &str,
tx: Sender<Self>,
inhibit_dlg: impl 'static + Fn(&Self) -> Option<Inhibit>
) -> Result<RawSignalCallback, Error>

Generate a signal handler function for GTK.

The returned function should convert the signals it revceives to the signal type, and transmit them over tx.

The signal must be one of the signals listed in SIGNALS.

Loading content...
Provided methods

fn connector() -> BuilderSingalConnector<Self, (), ()>

Entry point for connecting the signals supported by the BuilderSignal.

The BuilderSingalConnector returned by this method can be configured fluent-interface style and then either:

Get passed to connect_signals that connects signals emitted from a freshly created GTK builder to a freshly created Actix actor.

Get converted to a SignalRouter via route_to to connect GTK signals - from builders or otherwise - to an Actix actor.

Loading content...