Enum woab::Error

pub enum Error {

    IoError(Error),
    FromUtf8Error(FromUtf8Error),
    XmlError(Error),
    GtkBoolError(BoolError),
    WidgetMissingInBuilder(String),
    IncorrectWidgetTypeInBuilder {
        widget_id: String,
        expected_type: Type,
        actual_type: Type,
    },
    NoSuchSignalError(String),
    IncorrectSignalParameterType {
        signal: String,
        index: usize,
        expected_type: Type,
        actual_type: Type,
    },
    SignalParameterIndexOutOfBound {
        signal: String,
        index: usize,
        num_parameters: usize,
    },
    IncorrectEventParameter {
        signal: String,
        expected_type: &'static str,
        actual_type: EventType,
    },
    IncorrectActionParameter {
        signal: String,
        expected_type: VariantType,
        actual_type: VariantType,
    },
    NotAllParametersExtracted {
        signal: String,
        num_parameters: usize,
        num_extracted: usize,
    },
    WakerPerished(WakerPerished),

}

Variants

IoError(Error)

FromUtf8Error(FromUtf8Error)

XmlError(Error)

GtkBoolError(BoolError)

WidgetMissingInBuilder(String)

When extracting widgets using BuilderConnector::widgets and one of the widgets is missing.

IncorrectWidgetTypeInBuilder

Fields

widget_id: String

expected_type: Type

actual_type: Type

When extracting widgets using BuilderConnector::widgets and one of the widgets has the wrong type.

NoSuchSignalError(String)

When a signal handler does not recognize the name of the signal routed to it.

IncorrectSignalParameterType

Fields

signal: String

index: usize

expected_type: Type

actual_type: Type

When a signal parameter has the wrong type.

SignalParameterIndexOutOfBound

Fields

signal: String

index: usize

num_parameters: usize

When a signal has less parameters than what the handler expects.

IncorrectEventParameter

Fields

signal: String

expected_type: &'static str

actual_type: EventType

When an event signal’s parameter is of the the wrong event type.

IncorrectActionParameter

Fields

signal: String

expected_type: VariantType

actual_type: VariantType

When an action signal’s parameter is of the the wrong type.

NotAllParametersExtracted

Fields

signal: String

num_parameters: usize

num_extracted: usize

When a signal has more parameters than what the handler expects.

WakerPerished(WakerPerished)