Struct woab::Signal

pub struct Signal<T = ()>(_);

The generic signal WoAB uses.

The signal contains a name, list of parameters, and an optional tag. Route the signals from GTK to Actix actors using BuilderConnector, woab::route_signal or woab::route_action and handle them as actix messages, matching on their name and using woab::params! to get their parameters.

Implementations

impl<T: Clone> Signal<T>

pub fn creator(name: &str, tag: T) -> impl Fn(Vec<Value>) -> Self

impl<T> Signal<T>

pub fn new(name: Rc<String>, parameters: Vec<Value>, tag: T) -> Self

pub fn name(&self) -> &str

The name of the signal.

If the signal comes from a GTK builder, it is what’s written in the “Handler” field of the relevant signal in the “Signals” tabs of the Glade editor, or the handler attribute of the <signal> element in the XML.

If the signal comes from woab::route_signal, it is the third argument (actix_signal) passed to that function.

If the signal comes from woab::route_action, it is the name of the GIO action.

pub fn tag(&self) -> &T

The tag of the signal.

Tags are useful if the same GTK builder is instantiated many times, and all the instantiations are connected to the same actor. The actor can use the tag to identify the individual instantiation that sent the signal.

pub fn param<'a, P>(&'a self, index: usize) -> Result<P, Error> where
P: FromValue<'a>,
P: StaticType,

A parameter of the signal, converted to the appropriate type.

pub fn event_param<P: FromEvent + 'static>(&self) -> Result<P, Error>

The event parameter for event signals.

Convenience method - the parameter in event signals needs to be converted to gdk::Event first before it can be converted to its concrete type. This method runs both steps.

pub fn action_param<P: FromVariant>(&self) -> Result<P, Error>

The action parameter for stateless action signals, or the action state for stateful action signals.

Convenience method - the parameter in actions signals needs to be converted to glib::Variant first before it can be converted to its concrete type. This method runs both steps.

pub fn cant_handle(&self) -> SignalResult

An error indicating that an handler does not recognize the signal.

To be used in the catch-all arm of a signal match:

impl actix::Handler<woab::Signal> for MyActor {
    type Result = woab::SignalResult;

    fn handle(&mut self, msg: woab::Signal, _ctx: &mut <Self as actix::Actor>::Context) -> Self::Result {
        Ok(match msg.name() {
            "signal1" => {
                // ...
                None
            }
            "signal2" => {
                // ...
                None
            }
            "signal3" => {
                // ...
                None
            }
            _ => msg.cant_handle()?,
        })
    }
}

pub fn params<'a, R: SignalParamReceiver<'a>>(&'a self) -> Result<R, Error>

To be used with the woab::params! macro to extract all the signal’s parameters.